pyproject.toml
[project]
name = "ai-voice-agent-vapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.1",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "bcrypt>=4.0.0",
    "psycopg2-binary>=2.9.11",
    "psycopg[binary]>=3.3.2",
    "pydantic-settings>=2.12.0",
    "pydantic[email]>=2.12.5",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.21",
    "requests>=2.32.5",
    "uvicorn[standard]>=0.40.0",
    "vapi-server-sdk>=1.9.0",
]
